Card swipe actions P3: SwipeActionCard, anchored bubble, section snackbars (#889)
* SwipeActionCard, anchored ResolveBubbleMenu, and LocalSnackbarHostState (P3)

The reusable gesture component and the two shared-infrastructure changes
the tabs need, with component tests and no caller yet.

- SwipeActionCard in core/ui on Foundation's anchoredDraggable: both
  directions reveal the same action row, the open distance is the row's
  measured width, reveal-and-tap only, tap on an open card closes it.
  SwipeRevealController keeps one card open per list and closes on the
  first vertical scroll delta.
- ResolveBubbleMenu's tail now tracks the anchor's centre when the bubble
  is clamped to the window edge, and the bubble flips below the anchor when
  there is no room above. The placement math is a pure function with tests;
  the two edit forms are visually unchanged.
- LocalSnackbarHostState in core/ui/adaptive, provided by AdaptiveShellRoute
  around every section, so a quick action inside the shell can post a
  snackbar. ThingOverviewEvent.ShowError, which nothing consumed, becomes
  ShowMessage(UiText).
- First Compose UI tests in the repo: core/ui host tests run under
  Robolectric with ui-test-junit4 (six SwipeActionCard cases).

Closes #869
Closes #870
Closes #872
Closes #873
Closes #874


* Restore the ViewModel imports an optimize-imports pass stripped

UiText and the delete_failed accessor are used; the IDE could not resolve
the generated resource import and dropped all three. The iOS target caught
it. Also keeps the import reorder in AdaptiveShellRoute.


* Update code formatting

---------

+++ core/ui/src/commonMain/kotlin/dev/fanfly/wingslog/core/ui/brand/ThingHeroTimeline.kt
@@ -34,7 +34,8 @@ object ThingHeroTimeline {
   private const val FAN_OUT_MS = 700
   const val FAN_OUT_START =
     FAN_START + (GLYPH_COUNT - 1) * FAN_STAGGER + FAN_MS + FAN_HOLD_MS
-  const val FAN_END = FAN_OUT_START + (GLYPH_COUNT - 1) * FAN_STAGGER / 2 + FAN_OUT_MS
+  const val FAN_END =
+    FAN_OUT_START + (GLYPH_COUNT - 1) * FAN_STAGGER / 2 + FAN_OUT_MS
   const val IDLE_START = FAN_START + 540
   const val TOTAL_MS = FAN_END
 
@@ -138,11 +139,9 @@ object ThingHeroTimeline {
   fun idleWeight(ms: Int): Float = progress(ms, IDLE_START, 600)
 
   /** The plane's bob, matching the pre-existing hero: about six points up and a degree of roll. */
-  fun planeBobY(phase: Float): Float =
-    -0.017f * (0.5f + 0.5f * sin(phase).toFloat())
+  fun planeBobY(phase: Float): Float = -0.017f * (0.5f + 0.5f * sin(phase))
 
-  fun planeBobRotation(phase: Float): Float =
-    -0.25f + 1.25f * sin(phase).toFloat()
+  fun planeBobRotation(phase: Float): Float = -0.25f + 1.25f * sin(phase)
 
   const val IDLE_PERIOD_MS = 3400
   const val TWO_PI = (2 * PI).toFloat()
